Rename any_source Annotated to AnnotatedAny
Summary:
In D24463577, I move `general_type`, which represents either an `Annotated t` or `Inferred t` in the environment into `type.ml`. I move it there because I want to use that same variant structure in other data structures we in statement.ml, like the `Func_sig.t` data type's `return_t`.

The reason I want to track that information there is because we treat class properties like "functions" and treat annotations on the declaration like the "return type of a function."

By changing return_t to be a variant of Annotated / Inferred, we also can get rid of the `return_annot` parameter on `mk_func_sig`.

So where does this diff fit in? It's annoying to have multiple variants with the same name exported from the same file. To make it easier to write code down the line, I'm renaming the `Annotated` case in `any_source` to `AnnotatedAny`. I do the same for trust in the next diff.

Reviewed By: panagosg7

Differential Revision: D24463576

fbshipit-source-id: 1c306f9309afe9291e703492fcd5d8929e0fcc33