Introduce new shared heap API
Summary:
The new heap API makes it possible to allocate heap objects that contain
pointers to other heap objects. Follow-up work will use this API to write type
sigs into the shared heap, enabling them to be read lazily and on demand.

This diff only provides an API, but the API is unused. I would suggest reviewing
this first in terms of API design, with an eye towards the safety guarantees of
the public API.

The current sharedmem garbage collector is not sufficient to collect data stored
in this format, so the next bit of work will be to update the GC. Once this is
in place, we will be able to start storing type sigs in this format.

Reviewed By: nmote

Differential Revision: D24374023

fbshipit-source-id: e0039adaf630c90338078856a7e605eb85d05e0a