Keep track of annotation in func_sig instead of passing return_annot directly
Summary:
Instead of passing a `return_annot`, we can encode the information about whether an annotation is present or not directly on the Func_sig.t datatype. We use the `Type.annotated_or_inferred` type introduced in the previous diff to encode this information, the same way we do in the environment for variable bindings.

Since we treat class properties like getters while we analyze them, this will make it easy to push down the annotation on a method to the initializer.

Reviewed By: panagosg7

Differential Revision: D24463578

fbshipit-source-id: e465e312f692db4168070d059f886d61074730d3