Refactor sharedmem API: create prefix automatically
Summary:
To create the illusion of separate heaps, we construct hash keys by combining
the user key string with a unique prefix. Currently this prefix must be created
by users of the API, passed in the `Value` argument to the `NoCache` and
`WithCache` functors.

We only need the prefix value in the `Key` module, created by `KeyFunctor`. We
only create one `Key` module per `NoCache` (and one `NoCache` per `WithCache`),
so it suffices to simply create our prefix as part of `KeyFunctor`, dropping the
requirement from callers and simplifying the public API.

Reviewed By: jbrown215

Differential Revision: D24572000

fbshipit-source-id: 0034ce9b6c94844a33b18b1f4129c1b678777459