[easy] Bump xcode version in CircleCI config
Summary:
MacOS CircleCI builds have been broken for the last week, erroring that `failed to create host: Image xcode:9.0 is not supproted`. It looks like CircleCI deprecated the version of xcode we have configured (9.0) a while ago (https://discuss.circleci.com/t/removing-support-for-xcode-versions-9-0-0-9-3-0-and-9-4-0/30084), and builds are now breaking.

Let's bump to a non-deprecated version of xcode for CircleCI to fix macos builds. From https://circleci.com/docs/2.0/testing-ios/ it looks like we can upgrade to version 9.4.1 for a non-deprecated version of xcode 9.

Reviewed By: mroch

Differential Revision: D24607266

fbshipit-source-id: 27ba0200a03ab591f6eaa5bb11bc414a4535711e