[codemod] restore comment spacing after ocamlformat
Summary:
This restores most comments that had shifted vertically with the ocamlformat codemods.

For example comments of the form
```
        (* Copy the named exports from a source module into a target module. Used
        to implement `export * from 'SomeModule'`, with the current module as
        the target and the imported module as the source. *)
```
become
```
        (* Copy the named exports from a source module into a target module. Used
           to implement `export * from 'SomeModule'`, with the current module as
           the target and the imported module as the source. *)
```

To preserve meaningful internal indentation in comments like
```
(* Abc
      Def
      Ghi
        - Jkl
        - Mno
*)
```
the script records the indentation level of 'Def', and uses that as a prefix (positive
or negative) to align the rest of the lines of the comment. Ie. we want to preserve
the indentation of lines '- Jkl' and '- Mno', instead of flushing them as well to
the indentation level of 'Abc'. The output would look like:
```
(* Abc
   Def
   Ghi
     - Jkl
     - Mno
*)
```
see for example flow/src/typing/statement.ml

Reviewed By: samwgoldman

Differential Revision: D24613425

fbshipit-source-id: ecf940260b1c282cf832529c49016f62b340c13b