Simplify shared hashtbl API and separate from serialized heap API
Summary:
Before this change, the implementation of SharedMem.NoCache was very confusing,
split between separate New and Old modules, which shared a Raw module that
actually called into C.

Also, as of D22353111 (https://github.com/facebook/flow/commit/68544ca2f3386121640c4888a120435951b5fbc2), we always write to the heap before attempting to add the
heap address to the hash table. The New/Old/Raw API mixed these two operations.
For types-first 2.0, we want to be able to reuse the segmented hash table
abstraction by itself, without also pulling in the logic to
serialize/deserialize OCaml values into the heap.

This diff replaces the New/Old/Raw API in a way that is simpler (IMHO) and also
exposes the segmented hash table abstraction separately from any serialization
logic.

Instead of New/Old functors, we have a SegmentedHashtbl functor, which takes a
Key argument and returns a module that can be used to add, get, oldify, revive,
and remove hash table entries. The trick of adding a unique prefix is baked in.

Unlike the New/Old functors, the SegmentedHashtbl functor expects an address to
a heap object, not an OCaml value. That is, callers must first write to the heap
to get an address before they can manipulate the heap. For this, which was
previously part of the Raw API, I just moved this logic into NoCache directly.

Reviewed By: nmote

Differential Revision: D24594657

fbshipit-source-id: 202d667fcb51b9adc20d04eff6cbd1d8f532944e