Add hash table API for new heap
Summary:
The NoCache/WithCache functors were designed to work with the serialized heap.
That is, you parameterize them with a value type, and it will automatically
marshal that object and compress the resulting bytes when writing, and
decompress and deserialize when reading.

When using the new heap, we write to the heap in a separate step. If we simply
used the NoCache module with an address type as the value argument, the address
would be serialized, compressed, and stored in the heap. This is not what we
want.

We need the address to be stored in the hash table directly, because the GC
needs to be able to update addresses when values move. If we stored serialized
addresses in the heap, it would become invalid when the referred-to object moved
in the heap.

This diff builds on D24594657 to provide NoCacheAddr, a variant which does not
deal with serialization.

Reviewed By: nmote

Differential Revision: D24595217

fbshipit-source-id: d4708e741c2483b400bd6e1da3ce07415bddab58