upgrade xcode
Summary:
upgrades to the latest stable xcode available on Circle, but specifies that we should still target macOS 10.13. this lets us use a version of xcode that will be available as long as currently possible.

note: when 10.16 Big Sur is released soon, 10.13 will be de facto EOL and we can bump this to 10.14.

```
% otool -l bin/flow | grep -B1 -A3 LC_VERSION_MIN_MACOSX
Load command 9
      cmd LC_VERSION_MIN_MACOSX
  cmdsize 16
  version 10.13
      sdk 10.15.6
```

Pull Request resolved: https://github.com/facebook/flow/pull/8516

Reviewed By: Hans-Halverson

Differential Revision: D24622768

Pulled By: mroch

fbshipit-source-id: 97c50170f4476207198781ad48bbaf2b289f3e9e