mk_type_annotation returns annotated_or_inferred instead of Type.t
Summary:
There are a few places where the local inference code checks to see if a node is annotated. This case is already checked in the `mk_type_annotation` code, so instead of checking again we can just encode that information in the return value via `annotated_or_inferred`.

This diff changes the type but doesn't try to fix up any downstream sites. The next diff fixes up a few sites.

Reviewed By: panagosg7

Differential Revision: D24547730

fbshipit-source-id: cb5af8c5115d17b42a5fe4e7fd2a1e1316265333