allow setting stderr and log file verbosity independently
Summary:
`Hh_logger` can write both to stderr and also to a file, and the user can configure which verbosity level to log to both.

however, there are occasions where we might want to log different verbosities to each.

for example, I want to make the `flow lsp` process avoid writing to stderr (it should send errors over the LSP via stdout), but still capture those logs to a file.

Reviewed By: vrama628

Differential Revision: D24675986

fbshipit-source-id: f8cded26ff173698575e6d7db4f1c2539b5bffbe