New sharedmem GC
Summary:
In D24374023 (https://github.com/facebook/flow/commit/a9080e068e3f7ed7620456c16b3694e68a19e81d), we introduce a "new heap" API which includes a key new ability to
shared memory: the ability to write a heap object that contains a direct pointer
to another heap object.

The introduction of pointers between heap entries makes the existing sharedmem
GC insufficient. The current GC is only able to update the pointers contained
in the hash table, but now we need to also update the pointers contained in the
heap.

This diff implements a new GC, still mark-compact, using Jonker's 1978 threading
algorithm. The algorithm is explained in some depth in the inline comments. You
can also find an excellent description of the algorithm in this excellent paper
from Benedikt Meurer:

https://benediktmeurer.de/files/fast-garbage-compaction-with-interior-pointers.pdf

Note that there is a significant issue with the GC as proposed in this diff,
which still needs to be addressed before we can actually store pointerful
representations in the heap: we need to fix the accounting of wasted heap space.

Currently, tracking wasted heap space is trivial. When we delete a heap object,
we simply increment a counter by the space occupied by that one heap object.
This works when a given heap object is only kept alive by the hash table, but in
the new heap, one object X might be kept alive only by object Y, so deleting Y
would cause size(X) + size(Y) to become wasted.

There are many possible ways to approach this issue, but for now it is left as
TODO (see inline comment in hh_remove). Since this diff is already intricate, I
am going to propose a solution implemented on top of these changes.

Reviewed By: nmote

Differential Revision: D24510686

fbshipit-source-id: c3bae8526a51899f20cc40d0ad03a3e726901ef5