fix poorly positioned errors on incompatibility between this annotations
Summary: Way back when in the olden days of 2019, calls to `rec_flow_t` took `use_op` as an optional parameter, with `unknown_use` as the default. This was codemodded to make these calls explicit, but in some cases `unknown_use`s were added to model the old behavior that probably should actually be `use_op`s. This is one of them.

Reviewed By: jbrown215

Differential Revision: D24306475

fbshipit-source-id: 5a819bb1d0e1cb5fb2e9196b20645eded9806a68