type checking for this annotations in functions
Summary: This implements type checking for `this` annotations in functions, declared functions and function types. It does not implement any checking for `this` annotations in methods; that will come later. It is also important to note that this does not change the behavior of typechecking for functions where `this` is not annotated.

Reviewed By: panagosg7

Differential Revision: D24305945

fbshipit-source-id: 33a692544c56109d870187ac6e701c12a0102a60