fix hover and go-to-definition at end of token
Summary:
hover, go-to-definition and signature help all request info at a specific cursor location, which we then match against ranges from the AST.

we used `Loc.make file line col` for this, which produces a Loc covering `line:col` - `line:col + 1` --> the character at `line:col`.

when the cursor is at the end of a token, this is entirely outside the token, causing it not to match. `Loc.contains` checks whether `cursor._end.column <= token_loc._end.column`, which is false due to the `+ 1`.

we actually want a zero-width cursor covering no characters: `Loc.cursor`. that way, `Loc.contains` and `Reason.in_range` will include the cursor

Reviewed By: vrama628

Differential Revision: D24804613 (https://github.com/facebook/flow/commit/bbd7c3c69c16d11d9e2ef492795774ccf53e670b)

fbshipit-source-id: c9e5c16d4bf7d35f8e740bd6692e6d6ff2478959