Improve positioning of errors after speculation
Summary:
This is wading into some unfamiliar territory for me, but this diff attempts to improve error messages and fix an error in D24801184 whose primary location is a libdef. It does so by preserving use_op information in errors when the use_op is wrapped in a `Speculation` use_op, *unless* the error is itself a speculation error. The idea here is that speculative errors need to cut off the use op information since the overall speculation error contains a use op: otherwise, we'd get errors that look like:
```
Cannot cast X to T because:
   Either cannot cast X to T, because number [1] is incompatible with string [2], or
   Cannot cast X to T because ...
```
(the specifics here aren't important, the point is that the "cannot cast" message shouldn't be repeated.)

However, it seems like the way that `flow_error.ml` decides that the use_op should be omitted is incomplete: it omits error messages when the use op contains the `Speculation` use op, but that can appear inside a use op for an error that is not directly due to speculation, but happens after speculation finishes. (I think.)

The solution in this diff is to only omit the use op if the overall error being printed is a speculation error (i.e. is being created by a call to `mk_use_op_speculation_error`). Otherwise, we're in a situation where a non-speculation error includes the speculation use op, and we should show the full use op.

Reviewed By: panagosg7

Differential Revision: D24802912

fbshipit-source-id: 6d2a5e1b56eb9ffd71470a2604f1fa046bfc6d2d