fix autocomplete in JSX attribute values
Summary:
during autocomplete, we run typechecking with "hooks" that let us modify typechecking to account for the invalidness caused by the user still typing an incomplete token.

when you're typing an object literal's key, for example, we infer that property's value but don't add the property to the object because its key is bogus:

```
const x = { fo|: "a string" }
// x is inferred as an empty object since `fo` is not the real key, `foo` is
```

however, this was broken in JSX attributes. if the autocomplete token appeared anywhere within the attribute, we skipped typechecking the value at all, but still added the (potentially bogus) name to the element.

this diff brings JSX in line with object literals: we now infer the attribute's value like normal, and skip adding the attribute if the autocomplete cursor is within the attribute's name, not anywhere in the whole attribute.

the result is that completion inside attribute expressions now works properly.

```
(<div className={foo.ba|}></div>)
// before, skipped even looking at `foo.ba`
// after, recurses to find that `foo.ba` is a member expression completion
// and suggests properties of `foo` properly.
```

Reviewed By: nmote

Differential Revision: D25002179

fbshipit-source-id: 3a6ca23a84ff9b8b4bd0778ff6954b6f12f3bfc7