[new-generics] Fix for generic to keys flow
Summary:
This diff fixes an issue that arose for new generics when a flow like this occurs, `K ~> $Keys<{[K]: ty}>,` where `K` is a generic bounded by `string`. Prior to this diff, `K` would become `string` (its upper bound) before flowing into `$Keys<...>`, and from there the upper bound would be checked to see if `string` can flow into its keys type, `K`. Since `string ~> K` is an error, this leads to a problem in programs like this:

```
function foo<K: string>(
  key: K,
): $Keys<{[K]: any, ...}> {
  return key;
}
```
This diff modifies how `HasOwnPropT` (which is essentially a helper type for `KeysT` on the right) so that instead of just flowing `string` into the keys type of its incoming object and causing the above error, it contains the actual type to test against, which in this case is the `GenericT` `K`. With this change, the above code flows the `K` from `key` into the `K` of the return-type object, and the result is a success.

Reviewed By: panagosg7

Differential Revision: D25039995

fbshipit-source-id: 5e44febda874e16c515bb62a201c726457e7df20