fix in the order of processing predicate function constraints
Summary:
In examples like
```
function validate(x: ?string): ?string %checks {
  return (x === 'a' || x === 'b') ? x : null;
}
```
We check that the type of the return expression R is a subtype of ?T.
R is an OpenPredT that encodes the predicate expressed in the return
expression.

Before, we would process MaybeT, UnionT, etc. in the RHS *before*
processing OpenPredT on the LHS. This would break-up the upper bound
target prematurely:
```
OpenPredT ('a' | 'b' | null) ~> ?string
==>
OpenPredT('a' | 'b' | null) ~> string           // error
OpenPredT('a' | 'b' | null) ~> null | undefined // error
```

With this change, we process UnionT, MaybeT, etc. *after* "OpenPredT ~> _".
This allows
```
OpenPredT ('a' | 'b' | null) ~> ?string
```
to fall back to
```
'a' | 'b' | null ~> ?string // okay
```
since `?string` does not leverage the predicates within OpenPredT.

Reviewed By: samwgoldman

Differential Revision: D24967478

fbshipit-source-id: f5b95709ef655ecb0103f4daf60d76033c1dbf1b