[Parser] Use Literal node for init of enum boolean members in ESTree output
Summary:
Internally, Flow's AST stores the `init` for boolean enum members to be a `BooleanLiteral`, just like a `StringLiteral` is used for the init of  string enum members and a `NumberLiteral` is used for the init of number enum members. However in the ESTree output, both number and string members use a `Literal` node for the output, whereas boolean members use a raw boolean instead a `Literal` AST node.

Let's switch to also outputting a `Literal` node for boolean enum members in the ESTree output. This will be more consistent with number and string members as well as Flow's internal AST structure. Additionally this is already the AST format that Babel uses - Babel represents boolean enum member inits with a `BooleanLiteral` AST node. With this plus a change to Hermes, all three parsers will be aligned on the AST format for boolean enum member inits.

`ast-types` already specifies that boolean enum member inits can be either a raw boolean or a `Literal` node.

Prettier also already handles both raw booleans and `Literal` nodes here, as it works with the raw booleans produced by the Flow parser and the literal nodes produced by the Babel parser.

Reviewed By: gkz

Differential Revision: D25103039

fbshipit-source-id: 3029ca0f87106cdf456f99b36a652d68413fa5e5