[refactor] autocomplete text_edits -> text_edit
Summary:
an autocomplete result can return multiple text edits. for example, it obviously has to insert the completion, but it can also make other changes like inserting an `import`.

currently, the `text_edits` function takes a single edit and returns it as the entire list. instead, it should just return one so we can call it multiple times.

Reviewed By: vrama628

Differential Revision: D25122138

fbshipit-source-id: 0a17d6f8bcc03c09e0fff6661333d02bce770617