use record for type_sig type
Summary:
moves `Parsing_heaps.type_sig` to `Packed_type_sig.t` and turned it into a record.

I'm planning to pass this around to a bunch of recursive functions, where each one needs one or two parts of the tuple, and it is quite unwieldy to keep doing `let (_, _, _, foo, _, _) = type_sig in` in each function.

I moved it from `Parsing_heaps` to its own module in `type_sig` to avoid having to depend on `heap` unnecessarily.

Reviewed By: samwgoldman

Differential Revision: D25250927

fbshipit-source-id: e28709539bbc7f7f0770297cfb99a8f201b4c60b