Fix implicit instantiation infinite recursion with _
Summary:
Implicit instantiation has known methods for introducing non-termination. The comment in flow_js.ml just above the modifications made by this diff explain it. We introduced an implicit instantiation cache in order to cut constraint cycles that lead to non-termination.

Prior to `_`, there was no known way of introducing non-termination via explicitly instantiated type arguments in a function call. `_` is equivalent to implicit instantiation, which necessitates the use of the cache in the explicitly instantiated function calls as well. This diff adds that caching behavior and includes a test that hits a stack overflow when the caching mechanism is not present.

The implicit instantiation cache relies heavily on reasons. The changed test behavior is due exactly to a reason desc changing because it goes through the cache. Piping through the necessary information to keep that specific reason_desc around would keep the behavior as-is. However, that error message only pops up in a `classic` codebase that uses `_` in an input position, which is highly unlikely to appear now that types-first is the default. Given that this special behavior is not really needed anymore, I opted to remove it instead of keeping it around.

Reviewed By: panagosg7

Differential Revision: D25034431

fbshipit-source-id: baa99907eb130e6b624c81220512a4cd8ab12689