Add flag for post-inference implicit instantiation experimentation
Summary:
I plan to experiment with an implicit instantiation algorithm that pins down types instead of leaving unresolved OpenTs in the result. This is not possible during inference in Flow today and must take place in a post-inference pass, where we know that every annotation in a function signature is resolved and inspectable.

In order to allow myself to test my code and to avoid any performance complications that it may present, I am going to keep this experiment behind a flag.

The diffs in the rest of the stack will set up data structures and helper functions necessary to implement the implicit instantiation algorithm described by Pierce in his 00' paper.

Reviewed By: panagosg7

Differential Revision: D24734340

fbshipit-source-id: 9481e4c49eff0fc2c31a050f4047ddc1932eb61d