Use type_visitor to compute polarity of type parameters in function signatures
Summary:
We use the polarity-aware type_visitor to compute the polarity of type params in function signatures. There are a few interesting cases:
1. PolyTs may shadow the tparams bound at the function, so we need to make sure not to erroneously inspect those type params in signatures.
2. Some BoundTs are only used in upper bounds of other BoundTs. In order to compute their polarity, we visit those upper BoundTs each time we see the lower BoundT in a given polarity. It's possible that this is too strict-- i.e., if a BoundT is in a contravariant position, it is possible that we should not say that the upper BoundT is also in a contravariant position. This specific detail is something that can be worked out in a future diff after we see its effect on implicit instantiations.

I use the error constructor introduced earlier in the stack to print the polarities we calculate in the form of error messages to give reviewers more confidence in this approach.

Note that this diff only handles polymorphic function calls but not polymorphic constructors. I will add support for constructors in a later diff, but want to move on to the actual implicit instantiation algorithm before that.

Reviewed By: panagosg7

Differential Revision: D24734336

fbshipit-source-id: 0bbd0f09c9c26e6613d1ea16c3cc67d9a11abaf7