Introduce temporary error for implicit instantiation tests
Summary: There is no way to unit test type system code. In order to gain confidence in the implicit instantiation implementation, I'm introducing this error so that we can print results about the instantiation in error messages before the algorithm is complete. In the very next diff, I use this to print polarity information about the type parameters in an implicit instantiation. This error variant will be deleted when I can start writing full end-to-end tests of the implicit instantiation algorithm.

Reviewed By: panagosg7

Differential Revision: D24734335

fbshipit-source-id: f40981b16600c625f570f59fc0a529ec1ed8c5a8