pass AST to autocomplete
Summary:
autocomplete runs `Scope_builder`, which only works on untyped ASTs right now, but autocomplete also needs the typed AST.

previously, `check_contents` only returned the typed AST, so we had to strip the types off with `type_killer` for `Scope_builder`. but `check_contents` now returns the AST as well as the typed AST, so we can just pass it in instead.

still, in the future we should make `Scope_builder` work on a typed AST and not need both.

Reviewed By: vrama628

Differential Revision: D25306913

fbshipit-source-id: 01bfab2159ab9ed517d79c4d40faac0ac57aef93