fix autocomplete suggesting types in value positions
Summary:
added a flag to `Hoister` and `Scope_builder` to not add types to the scope, for cases that are only interested in values.

this fixes a bug with autocomplete in which we would suggest types in value positions (see new test; previously, it suggested all of those imported types!).

it also fixes a bug where autocomplete would list `import type`'d classes twice when completing types, because it was finding it once as a type (correctly) and again as a value (incorrectly).

Reviewed By: vrama628

Differential Revision: D25273237

fbshipit-source-id: 5f07002a739518c7edd830153ba0945610d09db6