extract entire token being autocompleted
Summary:
autocomplete is requested at a cursor location. we walk the AST and find the token under the cursor.

this diff returns the text of that token. it's currently not hooked up to anything, but is useful for filtering results based on what the user has typed so far.

for example, if you complete `fo|`, we now get back `Acid { name = "fo"; ... }`, so we can filter to only identifiers that (fuzzy) match `fo`, say `foo` and not `bar`. today, we return all identifiers that are in the cursor's scope, regardless of what you've typed.

Reviewed By: vrama628

Differential Revision: D25377140

fbshipit-source-id: d9823b2cec86326caaab90958d9173bf8fffd1b8