Fix issue when trying to remove non-existent heap entry
Summary:
hh_remove has an unenforced precondition that the removed item must exist.
Before D24594657 (https://github.com/facebook/flow/commit/6a35c176f16ecffb8a42a71ab8d4a7b508c7af88), this was satisfied because we would always check hh_mem before
calling hh_remove, but that diff removed the mem check.

If we oldify an entry, the slot remains occupied (the hash field is set) but the
addr is cleared to NULL. If the addr field is null, hh_remove would try to
dereference the NULL address, leading to bad behavior.

The NULL address is 0, which points to the hash slot of the first hash table
element. We read this 64 bit hash as a header and extract a size from it.
However, this size is just some meaningless number, and if the slot is occupied,
likely to be very large.

The net effect of this bug is that we added a large number to the "free heap
size" counter, which triggered GC too aggressively. I observed this behavior
by looking at the logs, observing that we triggered a compaction with only 0.6%
wasted space instead of the expected 20%.

Presumably, this also causes the hcounter_filled counter to be bogus. This is
scary as well, since the counter is unsigned, and would wrap around to MAX_LONG
if we decremented it past zero.

One question that remains is, why do we try to remove something after first
oldifying it? I traced the example I found which led to the "commit modules"
phase in the Module_js module.

If a change includes removing a file, we will first open a transaction including
that deleted file, which has the effect of oldifying the NameHeap entry for that
file's Haste module name. During commit modules, we also call
`remove_and_replace` which tries to directly remove the same NameHeap entry.

I suspect that the attempt to remove the entry is left over from an earlier
time, before cancelable rechecks and transactions, but I'm leaving further
investigation on this point for follow-up work.

Reviewed By: nmote

Differential Revision: D25387033

fbshipit-source-id: dc9684a840a0d41082c6cf1a7ca87c237acaf5c2