unique flowlib dir per user (fixes "Could not locate flowlib files")
Summary:
the files in `/tmp/flow` have default permissions of `644` (rw-r--r--). in v0.138.0, we started using a consistent directory for flowlibs, `/tmp/flow/flowlib_[hash]`, which ends up writable only by the user that first creates it. if multiple users run flow on the same machine, the second user gets a permission error.

to fix, we now include the effective user ID in the path: `/tmp/flow/flowlib_[hash]_[euid]`. the euid is what is used for file permission checks, so the directory is unique per user.

on Windows, the euid is always 1, but we write to `TEMP` which is per-user on Windows.

NOTE: technically, we still have the same problem with all the other files (logs, pid, sockets, etc) but we have things relying on those paths right now so I'm not changing them. it's much less likely for 2 users to run flow on the same directory. the workaround is to use `FLOW_TEMP_DIR` to write everything into a per-user temp directory, say `~/.cache/flow`.

Fixes https://github.com/facebook/flow/issues/8539

Reviewed By: samwgoldman

Differential Revision: D25400628 (https://github.com/facebook/flow/commit/35b1787691d95ae9cbb678e33853dd943a7be347)

fbshipit-source-id: c623853e17159f8a2bb660bf382a6b23d1d8540c