fix loc attached to namespace imports in file_sig
Summary: `File_sig` provides information about a module's imports, in which it includes information about the identifiers in the import statements. The location stored for `import * as foo from 'bar'` used to be the location corresponding to `* as foo`; this diff changes it to instead store the location corresponding to `foo`. The `File_sig` API provides this loc in a `type Flow_ast_utils.ident = Loc.t * string` where it'd be paired with the string `"foo"`, so the new behavior is more in line with expectations.

Reviewed By: nmote

Differential Revision: D24868725

fbshipit-source-id: 8d285f0add8dfe67463984d17fe099810dec86ea