Recompute the resolved requires hash when denormalizing saved state
Summary:
This hash has absolute file paths mixed in, so it needs to be recomputed instead of loaded from saved state.

This issue leads to Flow incorrectly believing that the resolved requires changed upon a recheck, since it makes that determination based on the hash alone. Currently, this doesn't cause any problems in practice, because this knowledge is only used:
* To decide whether to clear the direct dependents file cache, which would not be populated on the initial recheck for a file anyway.
* To decide whether we can consider directly-dependent files for skipping in the check phase, which we currently cannot do on the initial lazy recheck for a file anyway, because we don't have the sighashes.

However, my upcoming work will allow us to store sighashes in saved state and use them to skip work on initial lazy rechecks. This incorrect hash causes Flow to believe, incorrectly, that the resolved requires have changed, which stymies that work by preventing Flow from skipping direct dependent files in the check phase.

We could reduce the size of the saved state by removing this hash from it, now that we regenerate it anyway. I'll do that another time.

Reviewed By: samwgoldman

Differential Revision: D25414311

fbshipit-source-id: 1c08f357123a92c75ad4839d478f776c65becc32