Consolidate sharedmem collections
Summary:
To my knowledge, Flow has never been particularly disciplined with respect to
where sharedmem collections were attempted. We would sometimes attempt a
"gentle" collection during a recheck after removing something, and eventually an
"aggressive" collection during idle time.

With the introduction of cancelable rechecks, attempting collection during a
recheck became mostly useless. Existing entries are oldified at the beginning of
a recheck, but are not removed until the entire transaction commits.

I propose that there are exactly two places where it's necessary to attempt a
collection:

1. Between rechecks, during idle time, is the best time to try to collect. We
will have either committed or rolled back the transaction, in either case
potentially leaving garbage in the shared heap.

2. After a canceled recheck, before the coalesced re-attempt. This is really a
special-case of (1), since it's "between rechecks," but it does not hit the same
code paths. This is also not idle time, since the user is presumably waiting on
the follow-up recheck to complete.

This diff consolidates all calls to SharedMem_js.collect to those two places.

Reviewed By: nmote

Differential Revision: D25280252

fbshipit-source-id: 68d2d87bcaa513744576ea97abc6b5cc3377e9e3