Drop wrapping behavior from SharedMem_js module
Summary:
In the previous diff, I changed the SharedMem_js module. It used to trigger a
server status update, but that now happens in the newly introduced `on_compact`
hook.

After that change, it is no longer useful for SharedMem_js to "wrap" the
SharedMem module. Instead, we can export just the memory sampling utilities. In
a follow-up, I will move and rename this module to reflect that it contains only
these utilities.

A pattern lint ensured that we always used SharedMem_js instead of SharedMem, to
ensure we always had the wrapped behavior. This lint is no longer needed, so I
removed it here.

Reviewed By: nmote

Differential Revision: D25280253

fbshipit-source-id: 98cebc417ec504d1e73f1b51c372e9800ab3acd9