Add `on_compact` callback to sharedmem API
Summary:
The sharedmem API's `collect` function does not always compact the heap.
Instead, it checks the amount of wasted space and compacts conditionally.

Additionally, there are a few different hooks we'd like to trigger when the heap
is compacted:
1. Log the compaction event (Hh_logger and EventLogger)
2. Update the server status, because compaction can take a few seconds

Before this diff, these two hooks were triggered using two different methods:
1. Logging code was added directly to the implementation of `collect`, causing
the SharedMem module to be dependent on logging modules, with logging behaviors
hardcoded in.
2. The SharedMem_js module also wraps the SharedMem module, duplicates the
conditional logic (calls shared function). We need to be careful ensure we call
the wrapped module's `collect` function if we want the server status to be
updated. Internally we have a pattern lint for this.

This diff takes a different approach. We allow the application to provide an
explicit hook, by setting the `on_compact` ref on the SharedMem module. If
calling `collect` does ultimately trigger a compaction, the callback function
will be called.

The callback function has a slightly unusual signature, which allows it to
bracket the compaction. That is, the callback is applied with a single argument
before compaction, returning a continuation which is called afterward.

I have a few motivations for this change, but the primary motivation is to make
it easier to land the incremental GC. Similar to today, calling `collect` on the
incremental GC might not actually compact the heap, but the condition for when
is more complicated -- we need to have completed the marking and sweeping
phases, and have the appropriate of wasted space.

Otherwise, this change allows us to:
* Remove the `should_collect` function from the public API (done here)
* Remove the SharedMem_js wrapper module (next diff)

I'm not thrilled about the module-level ref for the callback. I considered some
alternatives, but ultimately decided they were not significant improvements.
That said, I'm certainly open to different approaches.

For example, I experimented passing the callback to the `collect` method
directly. This avoids the state in the module, but we call `collect` from two
places, and it's fragile to ensure they are both providing the callback.

I also experimented with initializing the `on_compact` callback as part of
`SharedMem.init` -- i.e., passing the callback to init, using it in subsequent
collect calls, but not exposing the ref itself. This still seems reasonable to
me, but it made testing code a bit more awkward.

Ultimately, the SharedMem module is awkwardly written, and is just a stateful
module by design, so I think the module-level ref is a good fit, but I would
welcome changes to make this module less stateful overall.

Reviewed By: nmote

Differential Revision: D25280250

fbshipit-source-id: 1f530e501c0e52d8f5510750bcd3ce160a64f570